requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==23.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
qrcode==8.2
pandas==2.3.3
openpyxl==3.1.5
reportlab==4.4.5
Pillow==11.3.0